Add `AutoUpdateAttribute` for identifying auto update properties
Auto update properties are like `UpdatedAt` and `CreatedAt`. If
a non versioned property is updated, then these auto update properties
might also be updated. If the only properties that have changed in an
update transaction are non versioned properties and auto update
properties, then do not create a new row.